Add logic to test if python is installed

+++ stable-diffusion-setup.sh
@@ -69,23 +69,39 @@ else
                 cd $sd_mount
                 git clone https://github.com/AUTOMATIC1111/stable-diffusion-webui.git 
             else
-                $priveleged_user mkdir /workspace/
-                cd $sd_mount
-                git clone https://github.com/AUTOMATIC1111/stable-diffusion-webui.git
-                # SD works with python3.8 but newer extensions like 3.10.  On Ubuntu that is a heavy refactor 
-                $priveleged_user apt-get install python3.8-venv
-                $priveleged_user apt-get install google-perftools libgoogle-perftools-dev
-                # Attempt ptyhon3.10 install 
-                $priveledge_user add-apt-repository ppa:deadsnakes/ppa
-                $priveledge_user apt-get update
-                $priveledge_user apt-get install python3.10
-                $priveledge_user update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.8 1
-                $priveledge_user update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.10 2
-                # Add code to fix pip now that python3.10 is installed
-                $priveledge_user apt remove --purge python3-apt
-
-                /workspace/stable-diffusion-webui/webui.sh
-            fi
+                if [ -d $sd_mount ]; then
+                    cd $sd_mount
+                    git clone https://github.com/AUTOMATIC1111/stable-diffusion-webui.git
+                    # SD works with python3.8 but newer extensions like 3.10.  On Ubuntu that is a heavy refactor 
+                    if command -v python3.8 &>/dev/null; then
+                        echo "Python 3.8 is already installed.  Continuing installation"
+                    else 
+                        $priveleged_user apt-get install python3.8-venv
+                        $priveleged_user apt-get install google-perftools libgoogle-perftools-dev
+                    fi
+                    if command -v python3.10 &>/dev/null; then
+                        echo "Python 3.10 is already installed.  Continuing installation"
+                    else 
+                    # Attempt ptyhon3.10 install 
+                        $priveledge_user add-apt-repository ppa:deadsnakes/ppa
+                        $priveledge_user apt-get update
+                        $priveledge_user apt-get install python3.10
+                        $priveledge_user update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.8 1
+                        $priveledge_user update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.10 2
+                        # Add code to fix pip now that python3.10 is installed
+                        $priveledge_user apt remove --purge python3-apt
+                        $priveledge_user apt autoclean
+                        $priveledge_user apt install python3-apt
+                        $priveledge_user apt install python3.10-distutils
+                        $priveledge_user python3.10 get-pip.py
+                        $priveledge_user apt install python3.10-venv
+                        # run webui.sh and cross fingers!
+                        if command -v pip --version &>dev/null; then
+                            echo "Upgrade to python3.10 failed"
+                        fi
+                fi
+                    # run webui.sh and cross fingers!
+                    $sd_mount/stable-diffusion-webui/webui.sh
 		;;
 		n|N|no|No)
             echo "exiting"
